Fix maven, make Common buildable
