+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/api/PlaceholderWrapper.java
