+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/api/VersionChecker.java
