+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/commands/DiscordCmd.java
