+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/controllers/ConfigurationController.java
