+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/controllers/VersionNotificationController.java
