+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/discord2mc/api/Discord2McConfigurationCache.java
