+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/discord2mc/listeners/MessageListener.java
