+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/eventlisteners/PlayerQuitEventListener.java
