+++ Bukkit/src/main/java/io/paradaux/hiberniadiscord/bukkit/events/ServerStopEvent.java
