+++ Bukkit/src/main/resources/HiberniaDiscord.properties
