+++ Bukkit/src/main/resources/locale.yml
