+++ Bukkit/src/main/resources/plugin.yml
@@ -1,0 +1,5 @@
+name: Bukkit
+version: ${project.version}
+main: io.paradaux.hiberniadiscord.bukkit.HiberniaDiscord
+api-version: 1.16
+description: HiberniaDiscord's Bukkit Functioanlity
