--- src/test/java/PlaceholdersTest.java
@@ -1,2 +1,0 @@
-public class PlaceholdersTest {
-}
