--- src/test/java/io/paradaux/hiberniadiscord/webhookutils/ServerWebhookTest.java
@@ -1,5 +1,0 @@
-package io.paradaux.hiberniadiscord.webhookutils;
-
-public class ServerWebhookTest {
-
-}
